+++ mlgw_bns/data_management.py
@@ -22,6 +22,7 @@ from typing import (
     Iterable,
     Iterator,
     Optional,
+    Tuple,
     Type,
     TypeVar,
     Union,
@@ -139,7 +140,7 @@ class SavableData:
 
 
 @dataclass
-class ParameterRanges(SavableData):
+class ParameterRanges:
     """Parameter ranges for waveform generation.
 
     The parameters should all be numpy arrays,
@@ -169,20 +170,12 @@ class ParameterRanges(SavableData):
 
     """
 
-    mass_range: tuple[float, float] = (2.0, 4.0)
-    q_range: tuple[float, float] = (1.0, 3.0)
-    lambda1_range: tuple[float, float] = (5.0, 5000.0)
-    lambda2_range: tuple[float, float] = (5.0, 5000.0)
-    chi1_range: tuple[float, float] = (-0.5, 0.5)
-    chi2_range: tuple[float, float] = (-0.5, 0.5)
-
-    group_name: ClassVar[str] = "parameter_ranges"
-
-    def __iter__(self) -> Iterator[Any]:
-        """Override this method for the purpose of saving
-        as an h5 file, which only works with arrays."""
-        for array_name in self._arrays_list():
-            yield np.array(getattr(self, array_name))
+    mass_range: Tuple[float, float] = (2.0, 4.0)
+    q_range: Tuple[float, float] = (1.0, 3.0)
+    lambda1_range: Tuple[float, float] = (5.0, 5000.0)
+    lambda2_range: Tuple[float, float] = (5.0, 5000.0)
+    chi1_range: Tuple[float, float] = (-0.5, 0.5)
+    chi2_range: Tuple[float, float] = (-0.5, 0.5)
 
     def check_parameters_in_ranges(self, params: ParametersWithExtrinsic) -> None:
         def within(x: float, x_range: tuple[float, float], name: str):